Updated scheduled-jobs to be a clusters command
