+++ croud/clusters/commands.py
@@ -27,14 +27,15 @@ from typing import Any, Dict, Optional, cast
 import bitmath
 import requests
 from tqdm.auto import tqdm
+from yarl import URL
 
 from croud.api import Client
 from croud.clusters.exceptions import AsyncOperationNotFound
-from croud.config import get_output_format
+from croud.config import CONFIG, get_output_format
 from croud.organizations.commands import op_upload_file_to_org
 from croud.printer import print_error, print_info, print_response, print_success
 from croud.tools.spinner import HALO
-from croud.util import require_confirmation
+from croud.util import grand_central_jwt_token, require_confirmation
 
 
 def clusters_get(args: Namespace) -> None:
@@ -797,6 +798,110 @@ def export_jobs_list(args: Namespace) -> None:
             "destination": _transform_export_job_destination,
         },
     )
+
+
+@grand_central_jwt_token
+def get_scheduled_jobs(args: Namespace) -> None:
+    client = _get_gc_client(args)
+
+    data, errors = client.get("/api/scheduled-jobs/")
+
+    print_response(
+        data=data,
+        errors=errors,
+        keys=["name", "id", "cron", "sql", "enabled", "next_run_time"],
+        output_fmt=get_output_format(args),
+    )
+
+    if errors or not data:
+        return
+
+
+@grand_central_jwt_token
+def get_scheduled_job_log(args: Namespace) -> None:
+    client = _get_gc_client(args)
+
+    data, errors = client.get(f"/api/scheduled-jobs/{args.job_id}/log")
+    print_response(
+        data=data,
+        errors=errors,
+        keys=["job_id", "start", "end", "error", "statements"],
+        output_fmt=get_output_format(args),
+    )
+
+    if errors or not data:
+        return
+
+
+@grand_central_jwt_token
+def create_scheduled_job(args: Namespace) -> None:
+    body = {
+        "name": args.name,
+        "cron": args.cron,
+        "sql": args.sql,
+        "enabled": args.enabled,
+    }
+
+    client = _get_gc_client(args)
+
+    data, errors = client.post("/api/scheduled-jobs/", body=body)
+    print_response(
+        data=data,
+        errors=errors,
+        keys=["name", "id", "cron", "sql", "enabled"],
+        output_fmt=get_output_format(args),
+    )
+
+    if errors or not data:
+        return
+
+
+@grand_central_jwt_token
+def delete_scheduled_job(args: Namespace) -> None:
+    client = _get_gc_client(args)
+
+    data, errors = client.delete(f"/api/scheduled-jobs/{args.job_id}")
+    print_response(
+        data=data,
+        errors=errors,
+        success_message="Scheduled job deleted.",
+        output_fmt=get_output_format(args),
+    )
+
+
+@grand_central_jwt_token
+def edit_scheduled_job(args: Namespace) -> None:
+    body = {
+        "name": args.name,
+        "cron": args.cron,
+        "sql": args.sql,
+        "enabled": args.enabled,
+    }
+
+    client = _get_gc_client(args)
+
+    data, errors = client.put(f"/api/scheduled-jobs/{args.job_id}", body=body)
+    print_response(
+        data=data,
+        errors=errors,
+        keys=["name", "id", "sql", "cron", "enabled"],
+        output_fmt=get_output_format(args),
+    )
+
+    if errors or not data:
+        return
+
+
+def _get_gc_client(args: Namespace) -> Client:
+    client = Client.from_args(args)
+    cluster, _ = client.get(f"/api/v2/clusters/{args.cluster_id}/")
+
+    url_region_cloud = cluster.get("fqdn").split(".", 1)[1][:-1]  # type: ignore
+    gc_url = f"https://{cluster.get('name')}.gc.{url_region_cloud}"  # type: ignore
+    client.base_url = URL(gc_url)
+    client.session.cookies.set("grand_central_session", CONFIG.gc_jwt_token)
+
+    return client
 
 
 def _transform_export_job_source(field):
