+++ docs/commands/scheduled-jobs.rst
@@ -13,25 +13,25 @@ The ``scheduled-jobs`` command lets you manage scheduled sql jobs for your clust
    :module: croud.__main__
    :func: get_parser
    :prog: croud
-   :path: scheduled-jobs
+   :path: clusters scheduled-jobs
    :nosubcommands:
 
 
-``scheduled-jobs create``
-=========================
+``clusters scheduled-jobs create``
+==================================
 
 .. argparse::
    :module: croud.__main__
    :func: get_parser
    :prog: croud
-   :path: scheduled-jobs create
+   :path: clusters scheduled-jobs create
 
 Example
 -------
 
 .. code-block:: console
 
-   sh$ croud scheduled-jobs create \
+   sh$ croud clusters scheduled-jobs create \
        --name test-job \
        --cluster-id 8d6a7c3c-61d5-11e9-a639-34e12d2331a1 \
        --cron "1 1 * * *" \
@@ -43,21 +43,21 @@ Example
    | test-job | 0EW7SX3ND87DY | 1 1 * * * | DELETE FROM test  | TRUE      |
    +----------+---------------+-----------+-------------------+-----------+
 
-``scheduled-jobs list``
-=======================
+``clusters scheduled-jobs list``
+================================
 
 .. argparse::
    :module: croud.__main__
    :func: get_parser
    :prog: croud
-   :path: scheduled-jobs list
+   :path: clusters scheduled-jobs list
 
 Example
 -------
 
 .. code-block:: console
 
-   sh$ croud scheduled-jobs list \
+   sh$ croud clusters scheduled-jobs list \
        --cluster-id 8d6a7c3c-61d5-11e9-a639-34e12d2331a1
    +----------+---------------+-----------+-------------------+-----------+---------------------------+
    | name     | id            | cron      | sql               | enabled   | next_run_time             |
@@ -65,21 +65,21 @@ Example
    | test-job | 0EW7SX3ND87DY | 1 1 * * * | DELETE FROM test  | TRUE      | 2024-01-20T01:01:00+00:00 |
    +----------+---------------+-----------+-------------------+-----------+---------------------------+
 
-``scheduled-jobs logs``
-=======================
+``clusters scheduled-jobs logs``
+================================
 
 .. argparse::
    :module: croud.__main__
    :func: get_parser
    :prog: croud
-   :path: scheduled-jobs logs
+   :path: clusters scheduled-jobs logs
 
 Example
 -------
 
 .. code-block:: console
 
-   sh$ croud scheduled-jobs logs \
+   sh$ croud clusters scheduled-jobs logs \
        --job-id 0EW7SX3ND87DY \
        --cluster-id 8d6a7c3c-61d5-11e9-a639-34e12d2331a1
    +---------------+----------------------------+----------------------------+---------+-----------------------------------------------------------------------+
@@ -88,21 +88,48 @@ Example
    | 0EW7SX3ND87DY | 2024-01-20T08:52:00.008000 | 2024-01-29T08:52:00.014000 | NULL    | {"0": {"duration": 0.0021747201681137085, "sql": "DELETE FROM test"}} |
    +---------------+----------------------------+----------------------------+---------+-----------------------------------------------------------------------+
 
-``scheduled-jobs delete``
-=========================
+``clusters scheduled-jobs edit``
+================================
 
 .. argparse::
    :module: croud.__main__
    :func: get_parser
    :prog: croud
-   :path: scheduled-jobs delete
+   :path: clusters scheduled-jobs edit
 
 Example
 -------
 
 .. code-block:: console
 
-   sh$ croud scheduled-jobs delete \
+   sh$ croud clusters scheduled-jobs edit \
+       --job-id 0EW7SX3ND87DY \
+       --cluster-id 8d6a7c3c-61d5-11e9-a639-34e12d2331a1 \
+       --name test-job-1 \
+       --cron "3 2 * * *" \
+       --sql "SELECT 1;" \
+       --enabled False
+   +------------+---------------+-----------+-----------+-----------+
+   | name       | id            | sql       | cron      | enabled   |
+   |------------+---------------+-----------+-----------+-----------|
+   | test-job-1 | 0EW7SX3ND87DY | SELECT 1; | 3 2 * * * | FALSE     |
+   +------------+---------------+-----------+-----------+-----------+
+
+``clusters scheduled-jobs delete``
+==================================
+
+.. argparse::
+   :module: croud.__main__
+   :func: get_parser
+   :prog: croud
+   :path: clusters scheduled-jobs delete
+
+Example
+-------
+
+.. code-block:: console
+
+   sh$ croud clusters scheduled-jobs delete \
        --job-id 0EW7SX3ND87DY
        --cluster-id 8d6a7c3c-61d5-11e9-a639-34e12d2331a1
    ==> Success: Scheduled job deleted.
